fetch lastest code

+++ plugins/kotlin/j2k/new/tests/testData/newJ2k/issues/kt-19353.kt
@@ -1,5 +1,5 @@
 class TestLongOfTwoInts {
     fun foo(x1: Int, x2: Int): Long {
-        return (x1 or (x2.toLong() shl 32).toInt()).toLong()
+        return x1.toLong() or (x2.toLong() shl 32)
     }
 }
